Refactor launch files
