+++ beluga_benchmark/docs/BENCHMARKING.md
@@ -11,16 +11,15 @@ This script will run the benchmark with the specified number of particles and re
 To run, use:
 
 ```bash
-ros2 run beluga_benchmark parameterized_run <NUMBER_OF_PARTICLES_EXPERIMENT_1> <NUMBER_OF_PARTICLES_EXPERIMENT_2> ... --initial-pose-y 2.0
+ros2 run beluga_benchmark parameterized_run <NUMBER_OF_PARTICLES_EXPERIMENT_1> <NUMBER_OF_PARTICLES_EXPERIMENT_2>
 ```
 
-Note that the initial pose value (x: 0, y: 2, theta: 0) is correct for the default rosbag. If not passed, particles will be initialized all around the map.
 The results of the different runs will be stored in folders named `benchmark_${N_PARTICLES}_particles_output`, where `N_PARTICLES` are the numbers specified in the above command.
 
 To run the same experiment using another AMCL node, e.g. nav2, use:
 
 ```bash
-ros2 run beluga_benchmark parameterized_run <NUMBER_OF_PARTICLES_EXPERIMENT_1> <NUMBER_OF_PARTICLES_EXPERIMENT_2> ... --initial-pose-y 2.0 --package nav2_amcl --executable amcl
+ros2 run beluga_benchmark parameterized_run <NUMBER_OF_PARTICLES_EXPERIMENT_1> <NUMBER_OF_PARTICLES_EXPERIMENT_2> --package nav2_amcl --executable amcl
 ```
 
 For other options, e.g. using a different rosbag, see:
